backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
dj-rest-auth==5.0.2
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.1
tweepy==4.14.0
dj-database-url==2.1.0
mysqlclient==2.2.1
python-decouple==3.8
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
django-storages==1.14.2
boto3==1.34.10
sentry-sdk==1.39.1
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==21.0.0

